Refactored jar e2e
