+++ tests/e2e_test.go
@@ -395,19 +395,6 @@ func TestSSH(t *testing.T) {
 	outputContains(t, out, uuid)
 }
 
-func TestJAR(t *testing.T) {
-	td := t.TempDir()
-	artifactPath := filepath.Join(td, "artifact.jar")
-
-	createSignedJar(t, artifactPath)
-
-	// If we do it twice, it should already exist
-	out := runCli(t, "upload", "--artifact", artifactPath, "--type", "jar")
-	outputContains(t, out, "Created entry at")
-	out = runCli(t, "upload", "--artifact", artifactPath, "--type", "jar")
-	outputContains(t, out, "Entry already exists")
-}
-
 func TestIntoto(t *testing.T) {
 	td := t.TempDir()
 	attestationPath := filepath.Join(td, "attestation.json")
@@ -642,102 +629,6 @@ func TestIntotoMultiSig(t *testing.T) {
 	out = runCli(t, "upload", "--artifact", attestationPath, "--type", "intoto", "--public-key", ecdsapubKeyPath, "--public-key", rsapubKeyPath)
 	outputContains(t, out, "Entry already exists")
 }
-
-/*
-func TestIntotoBlockV001(t *testing.T) {
-	td := t.TempDir()
-	attestationPath := filepath.Join(td, "attestation.json")
-	pubKeyPath := filepath.Join(td, "pub.pem")
-
-	// Get some random data so it's unique each run
-	d := randomData(t, 10)
-	id := base64.StdEncoding.EncodeToString(d)
-
-	it := in_toto.ProvenanceStatement{
-		StatementHeader: in_toto.StatementHeader{
-			Type:          in_toto.StatementInTotoV01,
-			PredicateType: slsa.PredicateSLSAProvenance,
-			Subject: []in_toto.Subject{
-				{
-					Name: "foobar",
-					Digest: slsa.DigestSet{
-						"foo": "bar",
-					},
-				},
-			},
-		},
-		Predicate: slsa.ProvenancePredicate{
-			Builder: slsa.ProvenanceBuilder{
-				ID: "foo" + id,
-			},
-		},
-	}
-
-	b, err := json.Marshal(it)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	pb, _ := pem.Decode([]byte(ecdsaPriv))
-	priv, err := x509.ParsePKCS8PrivateKey(pb.Bytes)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	s, err := signature.LoadECDSASigner(priv.(*ecdsa.PrivateKey), crypto.SHA256)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	signer, err := dsse.NewEnvelopeSigner(&verifier{
-		s: s,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	env, err := signer.SignPayload(in_toto.PayloadType, b)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	eb, err := json.Marshal(env)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	uaString := fmt.Sprintf("rekor-cli/%s (%s; %s)", version.GetVersionInfo().GitVersion, runtime.GOOS, runtime.GOARCH)
-
-	write(t, string(eb), attestationPath)
-	write(t, ecdsaPub, pubKeyPath)
-
-	rekorClient, err := client.GetRekorClient("http://localhost:3000", client.WithUserAgent(uaString))
-	if err != nil {
-		t.Fatal(err)
-	}
-	var entry models.ProposedEntry
-	params := entries.NewCreateLogEntryParams()
-	params.SetTimeout(time.Duration(30) * time.Second)
-
-	props := &types.ArtifactProperties{}
-
-	props.ArtifactPath = &url.URL{Path: attestationPath}
-
-	collectedKeys := []*url.URL{{Path: pubKeyPath}}
-	props.PublicKeyPaths = collectedKeys
-
-	entry, err = types.NewProposedEntry(context.Background(), "intoto", "0.0.1", *props)
-	if err != nil {
-		t.Fatal(err)
-	}
-	params.SetProposedEntry(entry)
-
-	_, err = rekorClient.Entries.CreateLogEntry(params)
-	if err != nil {
-		t.Fatalf("failed inserting v0.0.1 entry: %v", err)
-	}
-}
-*/
 
 func TestTimestampArtifact(t *testing.T) {
 	var out string
